NDMP: Switch ndmplib to 2015 data types.
- u_short ==> uint16_t
- u_long ==> uint32_t
- int ==> int32_t
- unsigned long long ==> uint64_t
- long long ==> int64_t
- unsigned long ==> uint32_t
- unsigned short ==> uint16_t

For GCC we now disable the following warning:

-Wunused-variable
-Wunused-but-set-variable
-Wformat
-Wenum-compare

For the Solaris Studio compiler we now disable the following warnings:

- E_ENUM_TYPE_MISMATCH_OP
- E_ENUM_TYPE_MISMATCH_ARG
- E_STATEMENT_NOT_REACHED

Those warnings are not too interesting for this code as
we are not going to fix those problems anyway.

The original problem we encountered was the following:

In solaris, the xdrrec_putlong function has the following check:

   if ((*lp > INT32_MAX) || (*lp < INT32_MIN))
      return (FALSE);

As the NDMP code assumes everywhere that long is 4 bytes, but on LP64 it
is 8 bytes, we get the problem that this check triggers and we get
always false back. The LP64 gets enabled when compiling your code for
64 bits on at least SPARC. We didn't see this phenomenon on 64 bits
compiled code on X86 with the Solaris Studio compiler on Illumos.

We fix this problem by using XDR_GETINT32 and XDR_PUTINT32 on LP64,
instead of XDR_GETLONG and XDR_PUTLONG which handles 4 bytes and so
works as desired.

+++ src/ndmp/ndma_comm_dispatch.c
@@ -1251,7 +1251,7 @@ ndmp_sxa_tape_mtio (struct ndm_session *sess,
 	ndmp9_error		error;
 	ndmp9_tape_mtio_op	tape_op;
 	int			will_write = 0;
-	unsigned long		resid = 0;
+	uint32_t		resid = 0;
 
       NDMS_WITH(ndmp9_tape_mtio)
 
@@ -1299,7 +1299,7 @@ ndmp_sxa_tape_write (struct ndm_session *sess,
   struct ndmp_xa_buf *xa, struct ndmconn *ref_conn)
 {
 	ndmp9_error		error;
-	unsigned long		done_count = 0;
+	uint32_t		done_count = 0;
 
       NDMS_WITH(ndmp9_tape_write)
 	if (request->data_out.data_out_len == 0) {
@@ -1349,7 +1349,7 @@ ndmp_sxa_tape_read (struct ndm_session *sess,
 {
 	struct ndm_tape_agent *	ta = sess->tape_acb;
 	ndmp9_error		error;
-	unsigned long		done_count = 0;
+	uint32_t		done_count = 0;
 
 	/*
 	 * We are about to read data into a tape buffer so make sure
@@ -2468,8 +2468,8 @@ ndmp_sxa_mover_set_window (struct ndm_session *sess,
 {
 	struct ndm_tape_agent *	ta = sess->tape_acb;
 	struct ndmp9_mover_get_state_reply *ms = &ta->mover_state;
-	unsigned long long	max_len;
-	unsigned long long	end_win;
+	uint64_t	max_len;
+	uint64_t	end_win;
 
       NDMS_WITH(ndmp9_mover_set_window)
 	ndmta_mover_sync_state (sess);
@@ -3237,7 +3237,7 @@ ndmp_sxa_fh_add_node (struct ndm_session *sess,
 
 int
 ndmta_local_mover_read (struct ndm_session *sess,
-  unsigned long long offset, unsigned long long length)
+  uint64_t offset, uint64_t length)
 {
 	struct ndm_tape_agent *	ta = sess->tape_acb;
 	struct ndmp9_mover_get_state_reply *ms = &ta->mover_state;
